converted all the functions to use templates_on.h and templates_off.h

+++ statistics/mean_source.c
@@ -1,4 +1,3 @@
-#include "source.h"
 
 double 
 FUNCTION(gsl_stats,mean) (const BASE data[], const size_t size)
@@ -6,7 +5,7 @@ FUNCTION(gsl_stats,mean) (const BASE data[], const size_t size)
   /* Compute the arithmetic mean of a dataset using the recurrence relation 
      mean_(n) = mean(n-1) + (data[n] - mean(n-1))/(n+1)   */
 
-  double mean = 0;
+  long double mean = 0;
   size_t i;
 
   for (i = 0; i < size; i++)
